ported over old (hopefully) scraper-compatible version of python_control
