--- src/ros/rover/python_control/python_control/controls/Direction.py
@@ -1,9 +1,0 @@
-#!/usr/bin/env python3
-
-from enum import Enum
-
-
-class Direction(Enum):
-    """Enum for the different directions of the motors"""
-    POSITIVE = 1
-    NEGATIVE = -1
